reorganize systemtest into separate files
also added a systemtest to verify that node status is restored on service restart
